e-commerce/requirements.txt
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt<4.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
stripe==11.5.0
redis==5.2.1
aioredis==2.0.1
sendgrid==6.11.0
slowapi==0.1.9
jinja2==3.1.5
python-multipart==0.0.20
pytest-html==4.0.2
celery==5.4.0
Pillow==11.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
chromadb==0.5.0
sentence-transformers==3.0.0
numpy==1.26.4
openai==1.54.0